requirements.txt
Flask==2.3.3
gunicorn==21.2.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Flask-Migrate==4.0.5

# Security packages
bcrypt==4.0.1
Werkzeug==3.0.1

# Production monitoring
Flask-Limiter==3.3.1
prometheus-flask-exporter==0.22.4

# Optional: Redis for caching and session storage
redis==4.6.0
Flask-Session==0.5.0